+++ src/Providers/CrudArrayValueProvider.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Crossview\Exphpress\Providers;
+
+
+class CrudArrayValueProvider extends WritableArrayValueProvider implements CrudProvider
+{
+	function unset( string $key ): CrudArrayValueProvider
+	{
+		unset( $this->values[ $key ] );
+		return $this;
+	}
+}
