+++ tests/Providers/HeadersProviderTest.php
@@ -13,22 +13,6 @@ class HeadersProviderTest extends TestCase
 		$this->assertEquals( 42, $provider->getRaw( 'the question' ) );
 	}
 
-	public function testUnsetRemovedHeaderFromValues(): void
-	{
-		$provider = new HeadersProvider( [ 'the question' => 42 ] );
-		$provider->unset( 'the question' );
-
-		$this->assertNull( $provider->get( 'the question' ) );
-	}
-
-	public function testUnsetReturnsHeadersProvider(): void
-	{
-		$provider = new HeadersProvider( [] );
-		$result   = $provider->unset( 'the question' ); // Doesn't throw because unset doesn't care if a symbol exists
-
-		$this->assertInstanceOf( HeadersProvider::class, $result );
-	}
-
 	/**
 	 * @runInSeparateProcess
 	 */
